refactor: notification 관련 속성 응집
