+++ src/context/notification/hook.ts
@@ -1,0 +1,12 @@
+import { useContext } from "react";
+import { NotificationContext } from ".";
+
+export const useNotificationContext = () => {
+  const context = useContext(NotificationContext);
+  if (context === undefined) {
+    throw new Error(
+      "useNotificationContext must be used within an NotificationProvider",
+    );
+  }
+  return context;
+};
